core: Finish internal modernization and test decomposition
Simplify subsystem build lists, remove superseded utilities, isolate backend implementations, and divide large test and scene sources into maintainable components.

+++ src/shaders/SoGLSLShaderObject.h
@@ -76,7 +76,7 @@ private:
   OBOL_GLhandle programHandle;
   OBOL_GLhandle shaderHandle;
   SbBool isattached;
-  int32_t programid;
+  uint64_t programid;
 };
 
 #endif /* ! OBOL_SOGLSLSHADEROBJECT_H */
